BIG changes. Moved all pages to their own files, and replaced polling with proper SSE handling!!!
