+++ lib/pages/advancedsearch_page.dart
@@ -1,0 +1,8 @@
+import 'package:flutter/material.dart';
+
+class AdvancedSearch extends StatelessWidget {
+  @override
+  Widget build(BuildContext context) {
+    return Center(child: Text("Advanced Search Coming Soon?"));
+  }
+}
